+++ pyproject.toml
@@ -57,8 +57,10 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.pytest.ini_options]
-#addopts = "--cov=src --cov-report=term-missing" # TODO uncommit
+addopts = "--cov=src --cov-report=term-missing" # TODO uncommit
 #asyncio_default_fixture_loop_scope = "session"
+#asyncio_mode = "auto"
+#asyncio_default_fixture_loop_scope = "function"
 pythonpath = "."
 filterwarnings = [
     "ignore:.*datetime\\.datetime\\.utcnow.*:DeprecationWarning"
